Redo 5 Sep 2 recs built with 2.2 & note on lengths
Appraised three more records as part
of issue #8. 😁

On General Makeposti, see corresponding Methods commit:

https://github.com/p1m-ortho/qs-global-ortho-search-queries/commit/a1e4df218780490f1b055a75a36bee5972d9883c

Note on lengths:

* When choosing a title for a record,
I follow journal guidelines on running
heads, which is usually 35–50 chars
or 5 words.
I use 35 chars AND 5 words.
This is for various purposes, so that
I don’t have to think if it fits in screen, etc.
* Let’s talk about other ‘lengths’ as well
if already brought up this topic.
For Highlights section of summaries,
I use 85-char limit as required
by Elsevier for their authors and by some
others.
That, too, is for the Highlights to fit in
tiny screenshots, and my experience
of two tweets showed that’s relevant.
* Finally, for the summary field in records
I will use somehow-standardy-looking
80 chars.
* All these lengths are subject to change;
also, I’m not even sure that’s really needed,
but noting this down for future reference.