Add next post link to t5 and increment edition
Did not increment the edition in some
other previous posts but am doing it here.
Not sure which strategy is better,
but will aim not to forget incrementing,
because the source file changes, so worth
incrementing anyway.

But, if forgot to increment here and there,
probably not a big deal.

Honestly, I need to automate it somehow.