Fix wording re: authors of hanula_evaluation_2024
Emily McDonald, who was the interviewee,
was not actually the _lead_ author but _one of_ the authors.
I did notice that she was the last among the authors,
likely signifying seniorship, but now I checked again with the article
and see that I did not use the accurate wording -
I apologize for it.