Revive twitter link and mention social media first pubs
I’ve been posting some content on SM recently,
which was not featured on the website.
I’ve been doing that for speed and community
engagement reasons - I often don’t have time
to write a full publication (or record a Stone research piece)
but have a minute to push it to socials.

Also, I’ve been back on Twitter recently bc
I’ve come out as a web3 enthusiast
and much of the web3 & crypto community
is on Twitter. I know that I need to replace
the icon with an X but don’t want to deal with that
at this time.